requirements.txt
apache-airflow
apache-kafka
pyspark
google-cloud-bigquery
pandas
